backend/requirements.txt
fastapi==0.78.0
python-dotenv
pandas
deta
uvicorn[standard]